python/requirements.txt
anthropic>=0.25.0
requests>=2.31.0
beautifulsoup4>=4.12.0
psycopg2-binary>=2.9.0
Pillow>=10.0.0
python-dotenv>=1.0.0
lxml>=4.9.0
reportlab>=4.0.0
